pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MANTIS"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bittensor",
    "botocore",
    "aiobotocore",
    "python-dotenv",
    "boto3",
    "ccxt",
    "numpy",
    "requests",
    "ipykernel",
    "torch",
    "pandas",
    "scikit-learn",
    "xgboost",
    "tensorflow",
    "yfinance",
    "matplotlib",
    "seaborn",
    "cloudflare>=0.7.0",
    "aiohttp",
    "xgboost",
    "tqdm",
]

[tool.setuptools]
# Use find_packages with exclusions to avoid auto-discovery of non-package directories
packages = {find = {exclude = ["logs*", "data*", "models*", "venv*", "test_storage*", "*.tests*", "*.test*"]}}

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yml", "*.yaml"]
